refactor: 💡 BDDAssertions 사용을 위해 테스트 코드 수정
✅ Closes: #11
