+++ src/main/java/com/postsquad/scoup/web/user/domain/User.java
